reorganising packages to allow import from outside of optimisation folder, and run the test __main__ in pyga.py too
